installing puppet on fedora/rhel now checks for links
When we were installing puppet on fedora and RHEL
  we get errors when trying to create a symlink for
  python-pip on second runs of install_puppet.
  This causes install_puppet to fail on those platforms
  because the script is configured to abort right away.

Check to see if the symlink exist before we
  try to create it.

+++ puppet/install_puppet.sh
@@ -85,7 +85,9 @@ function setup_puppet_fedora {
     # work-around is to just symlink pip-python to "fool" it.
     # See upstream issue:
     #  https://tickets.puppetlabs.com/browse/PUP-1082
-    ln -fs /usr/bin/pip /usr/bin/pip-python
+    if [ ! -L /usr/bin/pip-python ]; then
+        ln -fs /usr/bin/pip /usr/bin/pip-python
+    fi
 }
 
 function setup_puppet_rhel7 {
@@ -102,7 +104,9 @@ function setup_puppet_rhel7 {
     rpm -ivh $puppet_pkg
 
     # see comments in setup_puppet_fedora
-    ln -s /usr/bin/pip /usr/bin/pip-python
+    if [ ! -L /usr/bin/pip-python ]; then
+        ln -s /usr/bin/pip /usr/bin/pip-python
+    fi
 }
 
 function setup_puppet_rhel6 {
